+++ practice_with_YumemiWeather/Data/Request.swift
@@ -6,7 +6,6 @@
 //
 
 import Foundation
-import UIKit
 
 struct Request: Codable {
     var area: String
